requirements.txt
# AI Animation Studio Dependencies

# Core GUI Framework
PyQt6>=6.5.0
PyQt6-WebEngine>=6.5.0

# AI Generation
google-generativeai>=0.3.0

# Audio Processing (Optional)
librosa>=0.10.0
soundfile>=0.12.0

# Data Processing
numpy>=1.24.0

# Enhanced Features Dependencies
requests>=2.28.0

# Development Tools (Optional)
pytest>=7.0.0
black>=23.0.0
flake8>=6.0.0
